Add sub nav, basic structure, and github to rando
