+++ assets/global-styles.scss
@@ -10,6 +10,9 @@ $link-main-color--hover: rgb(20, 20, 20);
 $transparent-background-color: rgba(190, 190, 190, 0.3);
 $button--hover: rgb(60, 60, 60);
 $button--reset: rgb(196, 0, 0);
+$kinstone-green-color: #20a818;
+$kinstone-red-color: #a02818;
+$kinstone-blue-color: #2069c8;
 
 .container {
   display: flex;
