+++ package.json
@@ -25,7 +25,8 @@
     "nuxt": "^2.4.0",
     "nuxt-i18n": "^6.0.0",
     "vue-clickaway": "^2.2.2",
-    "vue-color": "^2.7.0"
+    "vue-color": "^2.7.0",
+    "vue-scrollto": "^2.18.1"
   },
   "devDependencies": {
     "@nuxtjs/eslint-config": "^0.0.1",
