+++ pages/save_editor.vue
@@ -2,9 +2,17 @@
   <main class="container">
     <img
       src="~/assets/Forest_Minish.png"
-      alt="Picture of Ezlo from Minish Cap" />
+      alt="Picture of a Forest Minish from Minish Cap" />
     <section class="sub-container">
-      Save Editor stuff
+      <h3 class="about">
+        About
+      </h3>
+      <h3 class="features">
+        Features
+      </h3>
+      <h3 class="download">
+        Download
+      </h3>
     </section>
   </main>
 </template>
